Use memory efficient versions of matrix/vector functions
Version 0.0.3 used 'functional' version of array operations that return
copies of their inputs.  This version uses destructive versions that
eliminate the need for memory allocation.  This results in a 5X
speedup on the 'stories15M' dataset.

Version 0.0.4 also removes the BLAS support.  This is a temporary
change until we can properly implement the vector/matrix operations in
a destructive manner in LLA, which is by default functional in design.